// src/app/models/user/role.model.ts
export class Role {
  id: any;
  name: string;
  orgCode: string;
  type: number;
  adGroup: string;
  iseSignAllowed: number;
  isInitalAllowed: number;
  parentRole: number;
  directorate: number;
  status:any;
}
export class RoleInfo {
  id: any;
  name: any;
  orgCode: any;
  type: any;
  status: any;
  adGroup: any;
  iseSignAllowed: any;
  isInitalAllowed: any;
  parentRole: any;
  pRoleStatus: any;
  directorate: any;
  pDirStatus: any;
}